fix(workspace): pin hickory=0.26.0-alpha.1, fix 6 pre-existing compile errors
- x3-dns-server: exact-pin all 4 hickory deps to =0.26.0-alpha.1 to prevent
  Cargo resolving to beta.1 (which uses rand 0.10 with broken API)
- parallel-proposer: add hex = 0.4.3 dependency (used in signature verifier)
- x3-wallet: fix u128/u32 type mismatch in health_factor calculation
- x3-gateway: annotate Ok::<(), GatewayError>(()) to resolve type ambiguity
- x3-dex: add sp-runtime dep; fix u128/u64 arithmetic; fix [u8;32]-to-u64 cast;
  fix u32*u64 in options; fix u32 negation in pool_analytics
- x3-atomic-client: add ClientT trait import + rpc_params! macro for request calls
- contention-predictor: add TransactionMeta struct; replace std::sync::Mutex
  with tokio::sync::Mutex; add Display for ModelType; fix f64 type annotations;
  fix &history borrow in retrain call

+++ crates/x3-dex/src/options.rs
@@ -150,7 +150,7 @@ impl OptionsEngine {
         };
 
         let gamma = (volatility_bps as u64 / 40).min(200);
-        let theta = -((volatility_bps * underlying_price / 10000) as i64 / 100);
+        let theta = -((volatility_bps as u64 * underlying_price / 10000) as i64 / 100);
         let vega = volatility_bps as u64 / 2;
         let rho = time_to_expiry_blocks as u64 / 100;
